Move json failures under `Json.Failures`
