+++ sbt-github/src/main/scala/com/alejandrohdezma/sbt/github/json/Json.scala
@@ -16,9 +16,8 @@
 
 package com.alejandrohdezma.sbt.github.json
 
+import scala.util.control.NoStackTrace
 import scala.util.parsing.combinator.JavaTokenParsers
-
-import com.alejandrohdezma.sbt.github.failure.NotAValidJSON
 
 object Json extends JavaTokenParsers {
 
@@ -26,7 +25,7 @@ object Json extends JavaTokenParsers {
 
   /** Parse the provided string into a [[Json.Value]] */
   def parse(s: String): Result[Json.Value] =
-    parseAll(`json-value`, s).map(Right(_)).getOrElse(Left(NotAValidJSON(s)))
+    parseAll(`json-value`, s).map(Right(_)).getOrElse(Left(Failures.NotAValidJSON(s)))
 
   @SuppressWarnings(Array("all"))
   private def `json-value`: Parser[Json.Value] = {
@@ -53,5 +52,21 @@ object Json extends JavaTokenParsers {
   case object False                                   extends Value
   case object True                                    extends Value
   case object Null                                    extends Value
+
+  object Failures {
+
+    final case class NotAJSONObject(value: Json.Value)
+        extends Throwable(s"is not a valid JSON object: $value")
+        with NoStackTrace
+
+    final case class NotAValidJSON(string: String)
+        extends Throwable(s"$string is not a valid JSON")
+        with NoStackTrace
+
+    final case class Path(value: String, fail: Throwable)
+        extends Throwable(s"$value => ${fail.getMessage}")
+        with NoStackTrace
+
+  }
 
 }
